+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - All workflows support environment selection via branch or manual dispatch
 - Streamline GitHub Actions workflows for consistency
 - Utility scripts work with dev and prod environments
+- User updated immediately when range deploy fails
 
 ## [0.3.6] - 2025-12-11